code for geminiAIjava

+++ src/main/java/com/WebVipers/gemini/service/GeminiApiService.java
@@ -7,10 +7,13 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.stereotype.Service;
 
+import com.WebVipers.gemini.model.BackStageResponseDTO;
+import com.WebVipers.gemini.model.BackstageRequestDTO;
 import com.google.genai.Client;
 import com.google.genai.types.Content;
 import com.google.genai.types.GenerateContentResponse;
@@ -19,25 +22,53 @@ import com.google.genai.types.Part;
 @Service
 public class GeminiApiService {
 
-	@Value("${root.directory}")
-	private String rootDirectory;
+	@Value("${root.directory.scafold.code}")
+	private String rootDirectoryScafoldCode;
+	
+	@Value("${root.directory.spec}")
+	private String rootDirectorySpec;
+	
 
-	public String getResponse(String prompt) throws Exception {
+	@Value("${backstage.url.location}")
+	private String backstageUrlLocation;
+
+	@Value("${backstage.catalog.path}")
+	private String backstageCatalogPath;
+
+	@Value("${git.spec.path}")
+	private String repositoryPath;
+
+	@Value("${git.remote.url.generated.code}")
+	private String remoteUrlGeneratedCode;
+	
+	@Value("${git.remote.url.spec}")
+	private String remoteUrlSpec;
+	
+	
+	@Autowired
+	private GitService gitService;
+
+	@Autowired
+	private WebClientService webClientService;
+
+	public String getScafoldResponse(String prompt) throws Exception {
 		Client client = null;
 		try {
 			client = new Client();
 
-			Content content = Content.fromParts(Part.fromText(generatePrompt(prompt)));
+			Content content = Content.fromParts(Part.fromText(generatePrompt(prompt, "Instructions.md")));
 
 			GenerateContentResponse response = client.models.generateContent("gemini-2.5-flash", content, null);
 			String finalResponse = response.text();
-			/*
-			 * ClassPathResource resource = new ClassPathResource("readme.txt"); String
-			 * finalResponse = null; try (InputStream inputStream =
-			 * resource.getInputStream()) { finalResponse = new
-			 * String(inputStream.readAllBytes(), StandardCharsets.UTF_8); }
-			 */
+
+			// ClassPathResource resource = new ClassPathResource("readme.txt");
+//			String finalResponse = null;
+//			try (InputStream inputStream = resource.getInputStream()) {
+//				finalResponse = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
+//			}
 			generateFiles(finalResponse);
+			gitService.commitAndPush(prompt, ".",rootDirectoryScafoldCode,remoteUrlGeneratedCode);
+			createServiceCatalog();
 			return finalResponse;
 		} catch (Exception e) {
 			throw e;
@@ -49,6 +80,39 @@ public class GeminiApiService {
 		}
 	}
 
+	public String getSpecResponse(String prompt) throws Exception {
+		Client client = null;
+		try {
+			client = new Client();
+			String finalResponse = null;
+			ClassPathResource resource = new ClassPathResource("Agent-template.md");
+			try (InputStream inputStream = resource.getInputStream()) {
+				Content content = Content.fromParts(Part.fromText(generatePrompt(prompt, "Instructions-spec.md")),
+						Part.fromBytes(inputStream.readAllBytes(), "text/plain"));
+				GenerateContentResponse response = client.models.generateContent("gemini-2.5-flash", content, null);
+				finalResponse = response.text();
+			}
+			String fileName="Agent-"+prompt.replaceAll(" ", "-")+".md";
+			writeContent(finalResponse, rootDirectorySpec,  fileName);
+			gitService.commitAndPush(prompt, ".",rootDirectorySpec,remoteUrlSpec);
+			return finalResponse;
+		} catch (Exception e) {
+			throw e;
+		} finally {
+			if (client != null) {
+				client.close();
+			}
+
+		}
+	}
+
+	public void createServiceCatalog() {
+
+		BackstageRequestDTO request = new BackstageRequestDTO("url", backstageCatalogPath);
+		BackStageResponseDTO response = webClientService.callPostApi(backstageUrlLocation, request,
+				BackStageResponseDTO.class);
+	}
+
 	public static void deleteDirectoryWithStream(Path directory) throws IOException {
 		if (!Files.exists(directory)) {
 			return;
@@ -58,7 +122,9 @@ public class GeminiApiService {
 			stream.sorted((a, b) -> b.compareTo(a)) // Reverse order to delete files before directories
 					.forEach(path -> {
 						try {
-							Files.delete(path);
+							if (!path.toString().contains(".git")
+									&& !path.toString().equalsIgnoreCase(directory.toString()))
+								Files.delete(path);
 						} catch (IOException e) {
 							throw new RuntimeException("Failed to delete: " + path, e);
 						}
@@ -66,8 +132,8 @@ public class GeminiApiService {
 		}
 	}
 
-	private String generatePrompt(String prompt) throws IOException {
-		ClassPathResource resource = new ClassPathResource("Instructions.md");
+	private String generatePrompt(String prompt, String instructionFile) throws IOException {
+		ClassPathResource resource = new ClassPathResource(instructionFile);
 		try (InputStream inputStream = resource.getInputStream()) {
 			String promptContent = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
 			promptContent = promptContent.replaceAll("\\{CHAT_BOT_CONTENT\\}", prompt);
@@ -76,15 +142,24 @@ public class GeminiApiService {
 	}
 
 	private Boolean generateFiles(String response) throws IOException {
-		deleteDirectoryWithStream(Path.of(rootDirectory));
-		writeContent(response, "", "README.md");
-		createFiles(Path.of(rootDirectory));
+		deleteDirectoryWithStream(Path.of(rootDirectoryScafoldCode));
+		writeContent(response, rootDirectoryScafoldCode, "README.md");
+		createFiles(Path.of(rootDirectoryScafoldCode));
+		addCatlogFile();
 		return true;
 	}
 
-	public void createFiles(Path filePath) throws IOException {
+	private void addCatlogFile() throws IOException {
+		ClassPathResource resource = new ClassPathResource("catalog-info.yaml");
+		try (InputStream inputStream = resource.getInputStream()) {
+			String content = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
+			writeContent(content, rootDirectoryScafoldCode, "catalog-info.yaml");
+		}
+	}
 
-		List<String> lines = Files.readAllLines(Path.of(rootDirectory + "README.md"));
+	private void createFiles(Path filePath) throws IOException {
+
+		List<String> lines = Files.readAllLines(Path.of(rootDirectoryScafoldCode + "README.md"));
 		String path = "";
 		String fileName = null;
 		StringBuffer buffer = new StringBuffer();
@@ -96,7 +171,7 @@ public class GeminiApiService {
 					path = "";
 				}
 				if (!path.endsWith("/")) {
-					path=path+"/";
+					path = path + "/";
 				}
 			}
 			if (line.startsWith("&&&&")) {
@@ -106,7 +181,7 @@ public class GeminiApiService {
 				readContent = readContent ? false : true;
 				if (!readContent) {
 					if (fileName != null)
-						writeContent(buffer.toString(), path.trim(), fileName.trim());
+						writeContent(buffer.toString(), rootDirectoryScafoldCode+path.trim(), fileName.trim());
 					path = "";
 					fileName = null;
 					buffer = new StringBuffer();
@@ -120,8 +195,7 @@ public class GeminiApiService {
 
 	}
 
-	public void writeContent(String content, String filePath, String filename) throws IOException {
-		String codePath = rootDirectory + filePath;
+	private void writeContent(String content, String codePath, String filename) throws IOException {
 		Files.createDirectories(Path.of(codePath));
 		String file = codePath + filename;
 		Files.createFile(Path.of(file));
